Rewrite appinject script
What this script does is that it installs backed-up apps from custom /sdcard location as user apps.
This will be especially useful after clean ROM installs (i always clean install ROMs almost at a weekly frequency).
I want to make it a flashable zip to reduce efforts even further. So now I have to write an updater-script required for the flashable zip.

Also, I want to try putting this script directly in the updater binary so that there's no need of even the updater script and additional files in the zip.

P.S. This script is not yet tested, may need changes (to the pm command mostly).